pyproject.toml
[project]
name = "smart-ai-resume"
version = "0.1.0"
description = "Executive Competitive Landscape Benchmark Engine — Multi-agent CV evaluation and repositioning system"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "DataOps Lab" }
]

dependencies = [
    "openai>=1.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-docx>=1.1.0",
    "pdfplumber>=0.11.0",
    "markdown>=3.6",
    "httpx>=0.27.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "typer>=0.12.0",
    "python-multipart>=0.0.9",
    "jinja2>=3.1.0",
    "beautifulsoup4>=4.12.0",
    "rich>=13.7.0",
    "streamlit>=1.45.0",
]

[project.optional-dependencies]
pdf-export = ["weasyprint>=62.0"]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]

[project.scripts]
smart-resume = "smart_resume.cli:app"

[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
disallow_untyped_defs = true
